+++ Windows API Projects/BouncingBallSimulation/BouncingBallSimulation/Circle.h
@@ -11,11 +11,11 @@ class Circle
 {
 public:
 	Circle();
-	Circle(float R, const Vec2& center);
+	Circle(double R, const Vec2& center);
 
 	bool hits(Circle& A, Vec2& normal);
 
-	float r; // radius
+	double r; // radius
 	Vec2  c; // center point
 };
 #endif // CIRCLE_H
