+++ Windows API Projects/BouncingBallSimulation/BouncingBallSimulation/Vec2.h
@@ -11,8 +11,8 @@ class Vec2
 {
 public:
 	Vec2();
-	Vec2(float s, float t);
-	Vec2(float d[2]);
+	Vec2(double s, double t);
+	Vec2(double d[2]);
 	Vec2(const POINT& p);
 
 	Vec2 operator+(const Vec2& rhs)const;
@@ -23,26 +23,26 @@ public:
 
 	void operator+=(const Vec2& rhs);
 	void operator-=(const Vec2& rhs);
-	void operator*=(float s);
-	void operator/=(float s);
+	void operator*=(double s);
+	void operator/=(double s);
 
-	float length();
+	double length();
 	Vec2& normalize();
 
-	float dot(const Vec2& rhs);
+	double dot(const Vec2& rhs);
 
 	Vec2& reflect(const Vec2& normal);
 
 	// Data members.
-	float x;
-	float y;
+	double x;
+	double y;
 };
 
 // Define with left hand and right hand sides reversed
 // so we can write both v*s and s*v.
-Vec2 operator*(const Vec2& v, float s);
-Vec2 operator*(float s, const Vec2& v);
-Vec2 operator/(const Vec2& v, float s);
-Vec2 operator/(float s, const Vec2& v);
+Vec2 operator*(const Vec2& v, double s);
+Vec2 operator*(double s, const Vec2& v);
+Vec2 operator/(const Vec2& v, double s);
+Vec2 operator/(double s, const Vec2& v);
 
 #endif // VEC2_H
